requirements.txt
ctcdecode==1.0.3
ds-ctcdecoder==0.9.3
g2p-en==2.1.0
h5py==3.1.0
jiwer==2.3.0
librosa==0.9.1
matplotlib==3.3.4
mir-eval==0.7
nltk==3.6.7
numpy==1.19.5
pandas==1.1.5
PyYAML==6.0
scipy==1.5.4
textdistance==4.2.2
torch==1.10.2+cu113
torch-scatter==2.0.9
torchaudio==0.10.2+cu113
tqdm==4.64.0